refactor(event): extract renderEventEditor/Viewer into shared module
Pulls the form markup, widget wiring, validation, and save/delete HTTP
helpers out of webpack/calendar-event-editor.js into a new
webpack/event-form.js module. The calendar modal becomes a thin shell
that owns only modal chrome (header/body/footer swaps, Bootstrap Modal
lifecycle, close/cleanup) and delegates field rendering to the module.

Behavior is preserved: same field set, same validation rules, same
TomSelect + Quill + datetime-local wiring, same save payload. The
`window.showEventForm` / `window.showNewEventForm` globals still work
identically for existing fullcalendar click handlers.

Prepares the ground for the full-page /event/editor/:id rewrite to
consume the exact same renderer (upcoming commits), so the two
surfaces can never drift apart on field naming, defaults, or
validation.
